+++ src/components/controls.ts
@@ -15,7 +15,7 @@ export function cellControlsTemplate(controls: ControlsDefinition) {
     return html`
         ${buttons.map((button) => 
             html`
-            <button @click="${button.callback}" class="cell-controls-button ${button.hide === undefined ? "auto-hide": button.hide} " title="${button.tooltip}">
+            <button @click="${button.callback}" class="btn cell-controls-button ${button.hide === undefined ? "auto-hide": button.hide} " title="${button.tooltip}">
                 ${button.icon({width: 20, height:20})}
             </button>
             `
@@ -28,7 +28,7 @@ export function getPropertiesIcons(cell: Cell, togglePropertyFunction: (name: st
     for(const prop of Object.getOwnPropertyNames(cell.metadata.properties)) {
         const propertyDef = registry.get(prop) || {icon: AlertCircleIcon, textEnabled: `Unknown property "${prop}"`, textDisabled: ``, name: `Unknown`};
         const templateResult = html`
-            <button @click=${() => togglePropertyFunction(prop)} class="cell-controls-button" title=${propertyDef.textEnabled}>
+            <button @click=${() => togglePropertyFunction(prop)} class="btn cell-controls-button" title=${propertyDef.textEnabled}>
                             ${propertyDef.icon({width: 15, height:15})}
             </button>
         `;
@@ -46,7 +46,7 @@ export function getPropertiesPopoverIcons(cell: Cell, togglePropertyFunction: (n
                 const helpText = isActive ? def.textEnabled : def.textDisabled;
                 const style = isActive ? "color: #8d27f4":"";
                 return html`
-                    <button style=${style} @click=${() => togglePropertyFunction(def.cellProperty)} class="cell-controls-button" title=${helpText}>
+                    <button style=${style} @click=${() => togglePropertyFunction(def.cellProperty)} class="btn cell-controls-button" title=${helpText}>
                                     ${def.icon({width: 16, height:16})}
                     </button>
                 `;